Fix undefined behavior in copy/moving ConfDataStoreDB and AllConfStoreDB after opening a file.
For ConfDataStoreDB, the database handle dbh_ keeps accessing the options_'s pointer after the opening. But the pointer will be invalid after the object is moved to another location, which can happens by a function returning a ConfDataStoreDB's object or during resizing a std::vector<ConfDataStoreDB>, for instance.

For AllConfStoreDB, allcfgs_.allconfigs can be freed during the lifetime of the object if the object is copied or moved.